+++ contracts/DiscordMarket.sol
@@ -6,7 +6,7 @@ import {AccessControlUpgradeable} from "@openzeppelin/contracts-upgradeable/acce
 import {IMarketBehavior} from "@devprotocol/protocol/contracts/interface/IMarketBehavior.sol";
 import {IMarket} from "@devprotocol/protocol/contracts/interface/IMarket.sol";
 
-contract YouTubeMarket is
+contract DiscordMarket is
 	IMarketBehavior,
 	PausableUpgradeable,
 	AccessControlUpgradeable
@@ -24,7 +24,7 @@ contract YouTubeMarket is
 	// event
 	event Registered(address _metrics, string _repository);
 	event Authenticated(string _repository, uint256 _status, string message);
-	event Query(string youtubeChannel, string publicSignature, address account);
+	event Query(string discordChannel, string publicSignature, address account);
 
 	function initialize() external initializer {
 		__AccessControl_init();
@@ -35,13 +35,13 @@ contract YouTubeMarket is
 	}
 
 	/*
-    _youtubeChannel: ex)
-                        YouTube channel id: UCN7m74tFgJJnoGL4zk6aJ6g
+    _discordChannel: ex)
+                        Discord channel id: UCN7m74tFgJJnoGL4zk6aJ6g
     _publicSignature: signature string(created by Khaos)
     */
 	function authenticate(
 		address _prop,
-		string memory _youtubeChannel,
+		string memory _discordChannel,
 		string memory _publicSignature,
 		string memory,
 		string memory,
@@ -51,8 +51,8 @@ contract YouTubeMarket is
 	) external override whenNotPaused returns (bool) {
 		require(msg.sender == associatedMarket, "invalid sender");
 
-		bytes32 key = createKey(_youtubeChannel);
-		emit Query(_youtubeChannel, _publicSignature, account);
+		bytes32 key = createKey(_discordChannel);
+		emit Query(_discordChannel, _publicSignature, account);
 		properties[key] = _prop;
 		markets[key] = _dest;
 		pendingAuthentication[key] = true;
@@ -60,7 +60,7 @@ contract YouTubeMarket is
 	}
 
 	function khaosCallback(
-		string memory _youtubeChannel,
+		string memory _discordChannel,
 		uint256 _status,
 		string memory _message
 	) external whenNotPaused {
@@ -70,10 +70,10 @@ contract YouTubeMarket is
 			"illegal access"
 		);
 		require(_status == 0, _message);
-		bytes32 key = createKey(_youtubeChannel);
+		bytes32 key = createKey(_discordChannel);
 		require(pendingAuthentication[key], "not while pending");
-		emit Authenticated(_youtubeChannel, _status, _message);
-		register(key, _youtubeChannel, properties[key]);
+		emit Authenticated(_discordChannel, _status, _message);
+		register(key, _discordChannel, properties[key]);
 	}
 
 	function register(
@@ -138,13 +138,13 @@ contract YouTubeMarket is
 	}
 
 	function name() external pure returns (string memory) {
-		return "YouTube";
+		return "Discord";
 	}
 
 	function schema() external pure override returns (string memory) {
 		return
 			// solhint-disable-next-line quotes
-			'["YouTube Channel (e.g, UCN7m74tFgJJnoGL4zk6aJ6g)", "Khaos Public Signature"]';
+			'["Discord guild id (e.g, 80351110224678912)", "Khaos Public Signature"]';
 	}
 
 	function pause() external onlyRole(DEFAULT_ADMIN_ROLE) whenNotPaused {
